Seed Flu programmes
When running the database seeds this ensures that some Flu sessions are
created so we can start to test the Flu programme. I've also reduced the
number of vaccines in the Flu programme factory by default (to improve
test performance) and picked a sample batch when generating vaccination
records (to improve realism).