Make available_delivery_sites explicit
This refactors the `Vaccine#available_delivery_sites` method to make the
list of available delivery sites explicit. I think this helps to improve
the readability of the code as you don't need to check on the list of
values in `VaccinationRecord`, and it means no new delivery sites will
be added accidentally.