Permit recording a Flu vaccination
This refactors the logic related to determining the common delivery
sites and the delivery method when recording a vaccination to allow for
Flu vaccinations to be recorded.

For now this doens't implement the full logic, specifically it doesn't
pick the delivery method according to the consent, but I've left a
`TODO` common so we know where to put this logic in the future.