Include Flu in reporting tests
This updates the tests for the three reports (CarePlus, Offline and
Mavis CSV) that output vaccination records to ensure that they work with
Flu.

For the most part Flu seems to work out of the box, but it's possible we
may need to add Flu-specific logic in the future. For now this gives us
some minimal test coverage on reporting.