Add UnsupportedProgramme exception
This adds an exception to capture the case where an unsupported
programme has been passed in to a method. This has no change on
behaviour as we already raise exceptions, but rather than raising a
`RuntimeError` we can raise a more specific class and handle the message
generation in one place.