Move mavis to bin
This moves the `mavis` command line tool to the `bin` directory where I
think it's better suited. Although it doesn't seem to be explicitly
documented anywhere, generally I think `bin` is for executables (of
which `mavis` is one, similar to `bin/rails` or `bin/good_job`) whereas
`script` contains project specific scripts that aren't executable on
their own (for that reason they have a file extension).