Save parent relationship when editing consent
This follows on from 46c2fa53e07f6b0b94fe396170873bdb390cbe59 which
didn't solve the original problem. In that commit only the check and
confirm page was updated to show the correct parent relationship, but
when we came to actually confirming the consent, the parent relationship
itself still wasn't changing.

This fixes that by replacing the use of `find_or_initialize_by` with a
simple `find` which ensures that the parent relationships remain
attached to the parent and therefore can be saved. I've also updated the
test to ensure this is working as expected and won't regress in the
future.

Jira-Issue: MAV-790