Start of follow up questions shouldn't trigger triage
When a parent fills out health questions, if they're answering the start
of a set of follow up questions, their answer to that question should
have no impact on whether the patient needs to be triaged or not.

Instead, only if any of the follow up health questions are responded to
with yes should the patient be triaged.