Create sessions per programme group
When first setting up an organisation, this ensures that a session is
created for each programme group (Flu, HPV and doubles) rather than
having a single session per school that covers all programmes.

This is necessary to support flu where there are significantly more
patients than the other programmes, and having a single session for each
school will mean lots of unnecessary patients showing for HPV and
doubles.

Jira-Issue: MAV-1293