Attach users to organisations when signing in
This makes it so that when a user signs in using CIS2, we attach their
user to the organisation they signed in to. This is useful when being
able to display a list of users associated with an organisation.

It doesn't handle the situation where a user leaves an organisation,
however such users will no longer be able to sign in, and it could be
useful to have a record of previous users.

Specifically, this change helps us towards two improvements:

- Fixing a bug where no nurses are displayed in the offline spreadsheet.
- Add the ability to choose which nurse performed pre-screening and
  identity checks when a health care assistant administers the vaccine.

Jira-Issue: MAV-1353