Create test patients and vaccination records
When testing the imms FHIR API code, it is useful to have patients in the system which can be used to test the API integration. These patients are limited because the integration environment of the API validates that the patients exist when a request is sent.

Create a script which will:
- read a cohort import
- Create patients from that import (skipping those whose DoBs are too early for Mavis)
- Create vaccination records for each patient created