Add drafting stage for session dates
The "back" button on the session dates edit page was broken. Even when
one clicked back, the changes made were still saved. MAV-516 includes a
bug report for this.

By adding a new model which is a draft version of the session dates we
can allow modification of session dates in the edit page without
immediately saving them to the database.