Add SyncVaccinationRecordToNHSEJob
This job will be responsible for syncing a Mavis vaccination record with
NHS England via the Immunisation API. The first task will be to simply
create the immunisation record if it hasn't already been synced,
implemented here. Funcionality will be added to update or delete the
record as necessary to keep the record in sync.

Jira-Issue: MAV-1477