Create Instruction model for Patient Specific Directions (PSD)
For patients requiring triage under the PSD protocol, nurses need
an option to attach a PSD instruction, which acts as a permission
for a healthcare assistant (HCA) to administer a specific vaccine
method at a given date/time.

We chose the name "Instruction" to avoid PSD in the name for better
abstraction, as the `protocol` column can have values like PSD,
PGD, National or Nurses Only.