feat(flu): Send reminders when vaccination/consent received last year
There was a bug that would prevent consent notifications from being sent
when the parent had already consented last year or when the child had
already been vaccinated last year. This behavior is incorrect for
seasonal vaccines for which consent is requested again every year and
which need to be taken again every year too.