Change pre-screening question to be directed to nurse
With flu it is a possibility that the child is not old enough to identify themselves. In which case they would also not
be old enough to confirm the pre-screening questions themselves. In that case it makes more sense to direct
the questions to the nurse who can check/confirm with the child accordingly.