fix: Change activity title to "Added to session ____"
Using the word "Invited" causes confusion for the user. It sounds like
an invitation has been sent to the child. In fact, the child has not yet
received any kind of invite at this point, they have only been added to
the session in MAVIS and will receive an invitation at the appropriate
time _if_ it is appropriate to invite them (e.g. not when they have
already had the vaccine or when they are deceased)

This confusion wording previous contributed to support task MAV-1226.