Fix flaky search form test
The test tries to create a vaccination record with a different programme
to that currently being search across, however because the vaccination
record factory picks programmes at random it's possible that sometimes
the two programmes will match. There's a unique index on the type of
programme meaning when two programmes of the same type are created we
get a unique constraint failure.