Add Session#eligible_programmes_for
This adds a new method on the session class which can be used to
determine the eligible programmes for a particular patient. This logic
already exists in the service but is duplicates in a number of places,
instead we can avoid that duplication and have the logic in a single
method.