Get programmes for GOV.UK Notify personalisation
At the moment, to determine the programmes (used in GOV.UK Notify
personalisation), they need to be passed in explicitly when we send the
email. This is causing an issue with the `vaccine_side_effects` variable
where it's not available for session reminders because the programmes
cannot be determined correctly.

Jira-Issue: MAV-1355