Use Patient#approved_vaccine_methods
This ensures that when sending emails that list vaccine side effects,
we're only include the side effects for the vaccines that the patient is
approved for. In most cases this has no effect, but for Flu it's
possible to be approved for either nasal or injection vaccines and
therefore we should only show the approriate vaccine methods.

Jira-Issue: MAV-1355