Add template ID for Flu administered email
This adds the template IDs for the email that we will send out after
Flu vaccination is administered.

This will also partially resolve this Sentry issue
(https://good-machine.sentry.io/issues/6670514347/) which is related to
the service trying to send emails for Flu where the templates don't
currently exist.

Jira-Issue: MAV-1529