Add consented_vaccine_methods_message personalisation
This adds a new variable which will be used in GOV.UK Notify templates
for flu where it includes the list of methods that the parents have
consented to. This is a variable as we need a fully customisable string
according to the consented methods and it's difficult to recreate this
purely using the Notify template format.

Jira-Issue: MAV-1503