Ensure flu is displayed lowercase in sentences
This adds a new method on the `Programme` model which can be used to get
the name of the programme approriate for rendering in the middle of a
sentence. For most programmes this will be the same as their name, but
for flu, when rendered in a sentence it should have a lowercase first
letter.

By having this method we can also start to replace the hard-coded
instances of `flu`.