Avoid hard-coded flu programme names
For flu, we often display the programme name with the "nasal spray" or
"injection" qualifier. Currently this is implemented each time the
string is displayed and usually using hard-coded strings. This commit
replaces the use of hard-coded strings with locale values and tries to
make the generation of this string more generic.

This has resulted in a few small content changes, but only to make the
content more consistent across the service.