Add talk_to_your_child_message personalisation
This replaces the `can_self_consent` variable that was added in
37c4891e152b7f4e3423ba192bc342059393dbc2 as we can't use the variable as
expected.

It was designed to be used with optional content, but the optional
content in this case contains a link which is not supported in optional
content. Instead we need to pass the entire block of content as a
variable (`talk_to_your_child_message` added in this commit) and that
handles the logic about whether to populate it or not.

Jira-Issue: MAV-1277