Ensure we have names for unused Notify templates
As we're removing or changing templates, we might have records of emails
and texts that were sent using a template that we no longer used. In
particular, to support the flu programme we've changed a large number of
templates in the previous commits.

Without keeping a track of the previous templates, users will start
seeing "Unknown Email" or "Unknown SMS" in the activity log.

I did think about updating the notify log entry model to track the
template names rather than the IDs, however this is a simpler solution.
Really we probably want to be storing the subject lines of the emails,
and that would be a future enhancement, but since that will require
further work in this area the simpler solution for now felt like the
easier way to move forward.