Prevent uploading flu vaccinations from previous years
This updates the importing process for the flu programme, preventing
files from being uploaded that contain records from previous academic
years.

This is likely to be temporary as we will want to allow users to upload
historical vaccinations, but only once we've made the necessary design
changes to support this properly.

Jira-Issue: MAV-1547