Don't ask full or half dose question by default
When recording a vaccination the default should be full dose, and the
nurse can change their answer to half dose if they need to. We shouldn't
be asking this question by default.

In fact, this question was only being shown when today's batch had been
selected due to the order of the questions, but now the behaviour is
consistent regardless of whether a default batch has been selected or
not.

Jira-Issue: MAV-1504