Update vaccine method filter to search on primary method
This updates the "Vaccine method" filter which users can use to look for
patients that have consented to a particular vaccine method, to filter
only on the "primary" method (which will be the first one in the array)
as this is the method that will be used in the UI and it's otherwise
confusing to see patients for the secondary methods where the content on
the page is based around the primary method.

In practice, this is only really relevant for nasal and injection where
a patient is primarily nasal and can optionally be injection.

Jira-Issue: MAV-1468