Filter on patients with approriate vaccine method
When displaying a list of patients who need vaccinating when filtering
on the vaccine method we should ensure that the vaccine method is
approriate for the vaccinations left to record. For example, if a
patient has consent for flu and HPV, and they've been given the nasal
spray vaccine for flu, if you now filter on "nasal spray" this patient
should no longer appear in the list as for HPV they're only eligible for
the injection.

Jira-Issue: MAV-1463