Show secondary consent vaccine methods
When displaying a consent either on the patient session page or when
viewing a consent directly, we should show if the parent also consented
to any alternative vaccine methods. This is only relevant for the flu
programme where you can consent to both the nasal spray and the
injection.

To support this I've also updated the summary list showing the consent
response to match the latest designs in the prototype.

Jira-Issue: MAV-1473