Don't ask follow up root questions in verbal consent
When recording health answers in verbal consent we don't want to ask for
any questions that don't require notes as these have no effect on triage
and instead we ask follow up questions which are the real questions.

This is needed to support the asthma question for the flu programme
where we ask about asthma, and then have some follow up questions. The
asthma question alone has no impact on the triage result.

Jira-Issue: MAV-1391