Move sync logic to immunisations_api module
There's more logic being added here and having it in a job can be
inconvenient, it makes more sense to have it grouped together in one
place and closer to the API calls. One way to describe it is that the
nhs_immunisations_api_* fields on the vaccination record are for the use
of the NHS::ImmunisationsAPI module, so it's responsible for managing
them.

Jira-Issue: MAV-1475