Fix bug in health question deduplicator
This fixes a bug in the deduplicator where users were encountering
infinite loops if the same question appeared in multiple vaccines as the
ordering wouldn't be set correctly.

What ended up happening is users would be taken back to the question
that would have already been answered in the first vaccine whereas
instead it should be skipped.