Create Location::ProgrammeYearGroup
This adds a new model which will be used to capture the year groups
administered for each programme and location. This is needed as we need
to support the ability for certain schools to administer vaccinations
beyond the normal year groups.

In reality, this should probably hang off a "programme-year" concept,
but we don't have one (yet) so for now it's based off the programme.