Add Programme#default_year_groups
This renames the existing `year_groups` method on programmes to instead
return the "default year groups" which is the same value, but by having
this name makes it clear that these year groups can be customised by
location.

The old method name `year_groups` still exists (aliased to the new one)
to support old code, but is deprecated and will be removed.