Use Location#create_default_programme_year_groups!
This ensures that when onboarding organisations or adding schools to an
organisation, we create a default set of programme year groups that's
suitable for use in the service.

I've also create a temporary Rake task
(`create_default_programme_year_groups`) that we'll need to use when
deploying this to production, but can be deleted once run.