Add patient after_update hook to sync vaccinations
There are a few places where a patient's NHS number may be updated, so
this attempts to catch them all and sync the patient's vaccination
records.

We could try to add explicit code to each spot where a patient's NHS
number is updated, however there are places such as ImportDuplicateForm,
where things get more complicated and having a hook such as this may be
preferable.