Allow sessions to not have registration
We need a way to disable registration on a per-session basis.
We might not build the UI to allow the users to do this themselves,
we can enable/disable
it in the backend.

The new boolean column `Session#requires_registration` will hide the
"Register" tab if false but still allow users to see all the patients
under "Record vaccinations".