Attach organisations to locations in tests
This ensures that when we create locations in the tests we attach them
to an organisation where approriate. Although so far the tests have
passed without this, we need this relationship to be set up (as it would
be in production) to be able to get the list of year groups administered
for a particular programme in a location.