Set vaccine method to `nil` if vaccinated
When updating the triage status of a patient, if the patient is
vaccinated their status moves from "safe to vaccinate" to "not
required". However, we were leaving the `vaccine_method` set to the
previous value meaning it was showing up in unexpected places in the
interface.

I think it makes sense to set the value back to `nil` to ensure it's in
sync with the overall triage status.

Jira-Issue: MAV-1500