Add academic year to programme controllers
This updates all the programme routes and controllers to include the
academic year allowing for the pages to be filtered according to the
academic year. This has no change to the current design as we will only
show the current academic year (as we do currently) but in the future it
allows us to show multiple academic years.