Use ConsentStatus in TriageStatus
This ensures that the logic used to determine the consent status is used
to determine the triage status. This removes some tech debt, and also
fixes a bug where if the consent is conflicting due to different
consented vaccine methods, triage is not marked as required. Currently
this is happening as the logic is duplicated incorrectly in both the
`ConsentStatus` and `TriageStatus`.

Jira-Issue: MAV-1591