Add academic_year column
This adds a new column to all three patient status models (consent,
triage and vaccination) as we now need to be able to filter status by
specific academic years in various views across the service.

In the follow up commits various parts of the service will be amended to
support this new column.