Ensure nurse can choose nasal or injection vaccination
This fixes a bug where nurses couldn't choose to record a nasal spray
vaccination if the patient had the option to be both vaccinated via
nasal spray or via injection.

This was happening because the `delivery_site` was coming through as an
array (for example `["nose", ""]`) because each top level radio button
on the form (for the vaccine method) had a conditional set of radio
buttons and these were coming through because "nose" was hard-coded as a
hidden value.

Instead this refactors how this action works to instead allow the
delivery site to be blank, and when blank, it checks that this is only
valid when the vaccine method (in this case nasal spray) has only one
delivery site available (in this case nose).

Jira-Issue: MAV-1338