Add BelongsToAcademicYear concern
This adds a new concern which can be shared across a number of different
models where the records should be in some way associated with an
academic year based on one of the date/time fields of the model.

By having a single concern we avoid having duplicate code across the
individual models.

Jira-Issue: MAV-267