Add number_of_preparation_days_before_academic_year_starts
This adds a settings variable which can be used to configure how many
days before an academic year starts that we enter the "preparation"
stage, which is where users can start to upload cohorts for the next
year.

We're not quite ready yet to enable this in production on the 1st
August, so once this variable is live we will edit the setting in
production to ensure it's closer to the 1st September.

Jira-Issue: MAV-1505