Add academic year preparation date
This adds support for adding academic years to the service before their
official start date through the introduce of a preparation period.

For now this has no impact on the service, but in the future this will
change which programme years are displayed on the programmes page and
which year groups patients are assigned to when importing records.

Jira-Issue: MAV-1505