Hide "Get a new consent response" button
This ensures that the button is hidden if no consent response is needed
for this patient to prevent nurses from accidentally recording a consent
response when not needed.

Specifically this prevents a scenario that was found to be possible in
testing, but wouldn't occur in reality where a patient already had
consent from a parent and then the nurse could request consent from
another parent which changed the status from consent given to consent
conflicting.

Jira-Issue: MAV-1591