Call create_default_programme_year_groups!
When first creating the generic clinic this ensures that the
`create_default_programme_year_groups!` method is called ensuring that
the generic clinic locations have the right year groups. This will only
have an impact on the tests currently because all organisations in
production already have their generic clinic locations created.

I'd actually like to refactor this out in to a separate class and call
it explicitly when organisations are onboarded rather than each time
the `generic_clinic` method is called, but that will come in a future
commit.