feat: filter consents by academic year
Consents from previous years are still considered for what parents to
inform in `session_notification.rb` as parents who gave consent last
year are still parents with known contact information.

Similarly, in new_or_existing_contact, parents who gave consent last year
are still used as suggested parents.

We also show all consents still on the parent's destroy page so that it
is known at the time of deletion that this historic information would be
lost.

The code style on academic_year.rb is ugly but this is need to make the
 linter pass