Ensure all changes are staged when recording offline
When recording offline, if a user is editing an existing record from the
spreadsheet, we should ensure that all changes to that record are staged
and highlighted to the user. This is to prevent the confusing scenario
where some of the changes are staged and some others are accepted
automatically.

Jira-Issue: MAV-1597