Move programme cohorts to overview page
This moves the cohorts cards that are currently displayed in the
"Cohorts" tab when looking at a programme in to the "Overview" tab,
matching the latest designs in the prototype.

When you select one of the cohort boxes, you are then taken to the
"Children" tab, with the year group filter pre-selected.

Jira-Issue: MAV-1509