Add AcademicYear#pending
This represents either the current academic year, or if during the
preparation for the next academic year, the next academic year.

Pending was chosen as the word to use to try and explain how it's
neither the `next` or the `current` academic year, and can be either at
different times in the year.