Ignore delivery site if there's only one
If a delivery site is selected, but the vaccine method means that
there's only one possible delivery site, we can ignore the provided
value.

This fixes a bug where if the user were to select injection and then
select a delivery site, but then go back and select nasal spray, the
delivery site for injection would still come through and it would be
invalid for nasal spray.

Jira-Issue: MAV-1656