Rename Team to Subteam
This renames the team model, and all associated models and foreign keys
to subteam. The functionality of the service before and after should be
the same.

We're doing this because we're going to be restructuring organisations
and teams in to three levels. What we currently call teams becomes
subteams, organisations becomes teams and we'll be introducing a new
top-level organisation to group teams together.

To support this change I've also added three new personalisation
variables to use in GOV.UK Notify: `subteam_name`, `subteam_email` and
`subteam_phone`. These are equivalent to the existing variables prefixed
with `team_` and once the templates have been updated we can remove
them.

Jira-Issue: MAV-1280