Remove flu performed by validation when uploading
We currently require the performed by first name and performed by last
name to be present when uploading flu records. Looking at the history of
the code, this has existed from the beginning well before flu was
properly supported and has stuck around.

Since we're not allowing historical flu records to be uploaded, we can
safely remove this validation as offline recording already has
validation to check the email address of who performed the vaccination
is provided.

Without this, it's currently not possible to upload flu records in
SystmOne format as the that format doesn't have a way of providing who
performed the vaccination.