Add PatientsAgedOutOfSchoolJob to schedule
This ensures that the job runs on a daily basis to ensure that patients
are correctly moved out of their school if approriate.

I've decided to schedule this daily rather than running it only on a
yearly basis to catch any uploads that might put the patient in the
wrong school during the preparation period for the next academic year,
to ensure that the job is well tested by running often, and to avoid the
need for a manual task to run each rollover.

Jira-Issue: MAV-1511