Set PGD protocol as a string
Rails enum types are returned as strings, whereas we were setting the
value to `:pgd` (a symbol). While this works correctly, Rails converts
the symbol in to a string, it flags the vaccination record as having
changed (`"pgd" to `:pgd`).

This only really affected offline recording where it's entirely possible
to upload details about an existing record where the UUID will ensure an
exact match. In that case, every record was showing as changed even when
they hadn't because the protocol value had changed from `"pgd"` to
`:pgd`.