Fix create_default_programme_year_groups task
We need to run this task after deploying 2.11 to production and it looks
like a refactor since the task was initially created and run in the test
environment has broken it. This fixes the Rake task to run correctly.

This task doesn't have any tests as it's a one off task that will be run
after a deploy and then deleted, it's been run in test and QA
successfully.