Configure number_of_preparation_days_before_academic_year_starts
This ensures that the right values are set in the various environments.
Specifically, in QA and test we want to give ourselves extra time for
testing the preparation period functionality, and in production we're
not ready for this feature to go live yet so we need to reduce the
number of days before it appears.