Ensure Location#programmes is ordered
This isn't used anywhere on the frontend so will have no user facing
changes, but it'll help to reduce flaky tests that rely on the ordering
of the programmes.

I felt that this change makes sense because if we were to start using
this in any user facing parts of the service, we would want the
programmes to be in the right order.