Simplify search form query parameters
This simplifies the query parameters for the patient search form to
avoid the `patient_search_form` prefix appearing for every parameter.
This is something Rails does by default and is useful for POST requests,
but with a GET request it can result in particularly long URLs which can
be difficult to work with. For example, what we have currently:

/patients?patient_search_form[q]=Thomas

Would be replaced with:

/patients?q=Thomas

One of the reasons to make this change is that I was concerned about
reaching the length limit of URLs, particularly on pages with lots of
filters.

Initially I wanted to implement this change by setting the `scope`
parameter to `form_with` to `nil`, however this results in it defaulting
to the `param_key` of the model. I tried a version where the
`model_name` was overridden to provide a custom `ActiveModel::Name`
class with a `nil` `param_key`, but this resulted in problems in other
areas where the model is used. Instead, the approach I've taken is to no
longer pass the search form model to the builder.