Import patients in to pending year group
When importing patient records this ensures that the year group is
determined using the pending academic year. This means that during the
preparation period for the upcoming academic year, patients can be
imported using the year group they will be rather than their current
year group.

In making this change I've been able to make the `academic_year`
argument on the `to_year_group` and `to_birth_academic_year` methods a
required argument, ensuring we're more delibrate about which academic
year we're using.