feat: Add academic_year column to consent form
The consent forms academic year should be determined by the session it
is meant for, not by its creation date. We do not want to store the
session that it belongs to so we must store the academic year it belongs
to at the time of creation of the form.