Rename in_programmes scope
To `appear_in_programmes` to make it clearer what this scope
represents. Initially I was going to go with `eligible_in_programmes`
but I wanted to avoid the word eligible as it's not quite correct.

For example, a patient may be eligible for a programme based on their
year group, but only exist in sessions where that programme isn't
administered at their year group, therefore they remain eligible, but
wouldn't show in those sessions for that programme.

Jira-Issue: MAV-1293
Jira-Issue: MAV-1429