Use radio buttons for organisation selection
This makes a small design change to the page where the user selects
their organisation in non-CIS2 environments in preparation for its use
in CIS2 environments. When the teams restructuring work is done, users
will log in under their organisation ODS code, and will be required to
pick a team.

A design hasn't been fully finished for this yet, other than radio
buttons rather than cards, so this changes the design to use radio
buttons.

Jira-Issue: MAV-1280