Set academic_year_today_override to nil
It turns out a blank string is not a valid parameter store value, so we
need to set it to a specific value which is understood to mean unset.

We can't remove this variable while it's unused because we need to be
able to change it dynamically, and therefore it needs to exist.