Record already had vaccinations from consent response
If the parents responded to consent during an academic year and
indicated that the patient was already vaccinated, provided that the
academic year has come to an end, we should be recording these as
vaccination records to ensure that the patient doesn't get asked for
consent again next year.

To implement this I've created a job which only runs during the
preparation period before an academic year starts.

Jira-Issue: MAV-1517