Import class lists in to schools
Rather than sessions as we have currently. This is needed because we now
support multiple sessions for the same school location.

Instead of importing a class list in to a specific session, we now
import the class list for a school, and the school moves handles adding
the patients to all the necessary sessions.

Jira-Issue: MAV-1293