Add option to not send consent refusal to parents
This adds a question when recording verbal consent, and where the parent
has refused for personal reasons, allowing the nurses to not send a
confirmation to this parent.

This will be used in the preparation period for the next academic year
where nurses will record refusal for parents they know will refuse
consent and wouldn't like to receive confirmation.

Jira-Issue: MAV-1586