Show eligible programmes on patient page
This adds a new section to the patient page showing which programmes the
patient is eligible for at which time. In the prototype, this card is
supposed to replace the cohorts card, but since we haven't built the
archive functionality that's designed to replace "Remove from cohort"
I've left it in place.

Jira-Issue: MAV-1516