Add URN/postcode to school options drop down
When schools have the same name, the autocomplete controller only finds the first match with
the name and renders all other options with the same name as a duplicate of the first. This means
the user will not be able to see or select the other options. Making the option names uniqe (e.g. by adding the URN)
ensures all the options render and can be selected. We only add the URN for SAIS view and postcode for parent
view when schools have the same name, not for all schools.