Refactor consent refusal reasons
This refactors how the consent refusal reasons are displayed when a user
is able to choose between them. There are three pages where the reasons
are displayed and they were duplicated in each one:

- Parent consent form
- Nurse consent form
- Withdrawing consent

I've introduced a helper method that can be used to get the approriate
options for a particular consent-like object. This also fixes two bugs:
one where the gelatine option wasn't showing on the withdraw form, and
another where gelatine wasn't selectable.

I've also taken the opportunity to reduce duplicate in the locale
strings, by using only the locale strings for the `Consent` model in
all places.

Jira-Issue: MAV-1696