Create previous years sessions when onboarding
When onboarding a test for use in the regression tests, we should create
sessions for the pending academic year, plus the previous academic year.

This ensures that during the preparation period, we have sessions that
are valid for the current academic year (so vaccinations can be tested)
and for the pending academic year.