Improve session ordering
This improves the ordering of sessions on the new sessions page.
Specifically, when you search for a location by name this ensures that
the results are ordered by the closest match, and if not searching by
name then the sessions are ordered by date with the next closest
sessions appearing at the top of the results.