Add VaccinationRecord#from_fhir_record
This function takes a FHIR record and converts it into a Mavis `VaccinationRecord`.

If the vaccine's SNOMED product code doesn't match any within Mavis (or a team isn't provided), then this code will add the product code, description, batch number and expiry to the notes of the record. This is an effort to retain the relevant information without needing to refactor a lot of code and add potentially needless `Vaccine` records.

If this happens, it raises a Sentry error, which will allow us to track if there are consistently codes which are missed. If this is the case, these records will be readily overwritten, so a rerun of the search job should overwrite with whatever new configuration we choose.