Refactor how distinct patients are determined
In aebbe50364e78c826b410f4fb7f521d4853ff1b8 we attempted to fix the
issue of duplicate patients being shown by using a `distinct` on the
scope. This works, but it causes issues if we want to start ordering or
grouping the patients.

Specifically, when using `DISTINCT` or `DISTINCT ON`, anything in the
`ORDER BY` needs to be included, and this makes this ricky to use as you
need to know before applying the ordering, what columns to select.

Instead, we can use a subquery to achieve the same result while avoiding
the issues around using `DISTINCT`.

This also fixes a bug where you can't search for children on the
programmes view.

Sentry-Issue: 6782169105