Make programme stats consistent
This fixes the programme stats on the index and overview pages to be
consistent.

We're not sure if these stats are the best, but at least they are now
consistent across the two pages. I haven't added any specific tests for
these as it's likely we will change these stats in the future.

The vaccintions count will be administered records and for all the
patients in the programme, not just ones administered this academic
year. The reason for that is because the number of patients includes
patients who are already vaccinated.