Add SelectOrganisationForm
This adds a new form which handles the select an organisation page,
ensuring that we validate that the user has selected a valid
organisation and showing an error message if not.

This fixes an issue where if you selected an invalid organisation or
didn't select an organisation, then the user would be taken to an error
page.