Show import issues count on header item
This updates the "Import" header item to show the count of the number of
import issues rather than the count of the number of imports. This was
actually supposed to be the case originally but it was built
incorrectly.

I've implemented this by using a helper method since we need this number
to be accessible in both the import issues navigation item and the
global header item. A helper means it's available to every page.

It's a little awkard to test the helper on its own as it relies on the
use of `policy_scope`, so instead test coverage is provided by the
feature tests.

Jira-Issue: MAV-1695