Fix tests to handle preparation period
During the preparation period there are two academic years in play, the
current academic year and the upcoming one. Various parts of the service
apply to different academc years during the preparation period.

By default, the prepartation period starts on the 1st August and
therefore all our tests are running during this period. This has
highlighted that a number of tests assume that we're not in the
preparation period and have started failing.

I've fixed these in most cases by freezing the time to before the
preparation period starts, which also works as a test of the preparation
period since the failure demonstrates that different logic is being
applied. However, as we continue to develop the preparation period it
would make sense to write tests both in and out of the period.

For now, this prevents PRs and deploys from being blocked.